+++ lib/service/Appointments.dart
@@ -14,8 +14,13 @@
  * limitations under the License.
  */
 
+import 'dart:io';
+
+import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/model/wellness/Appointment.dart';
+import 'package:path/path.dart';
+import 'package:path_provider/path_provider.dart';
 import 'package:rokwire_plugin/model/explore.dart';
 import 'package:rokwire_plugin/service/app_livecycle.dart';
 import 'package:rokwire_plugin/service/deep_link.dart';
@@ -30,12 +35,17 @@ import 'package:illinois/service/Auth2.dart';
 
 class Appointments with Service implements ExploreJsonHandler, NotificationsListener {
   static const String notifyAppointmentDetail = "edu.illinois.rokwire.appointments.detail";
-  static const String notifyAppointmentsAccountUpdated = "edu.illinois.rokwire.appointments.account.updated";
+  static const String notifyAppointmentsChanged = "edu.illinois.rokwire.appointments.changed";
+  static const String notifyAppointmentsAccountChanged = "edu.illinois.rokwire.appointments.account.changed";
   
   static const String _appointmentRemindersNotificationsEnabledKey = 'edu.illinois.rokwire.settings.inbox.notification.appointments.reminders.notifications.enabled';
   
   DateTime? _pausedDateTime;
+  late Directory _appDocDir;
+
+  static const String _appointmentsCacheDocName = "appointments.json";
   
+  List<Appointment>? _appointments;
   AppointmentsAccount? _account;
   bool? _isLastAccountResponseSuccessful;
 
@@ -68,6 +78,21 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
 
   @override
   Future<void> initService() async {
+    _appDocDir = await getApplicationDocumentsDirectory();
+
+    // Init appointments
+    _appointments = await _loadAppointmentsFromCache();
+    if (_appointments != null) {
+      _updateAppointments();
+    } else {
+      String? appointmentsJsonString = await _loadAppointmentsStringFromNet();
+      _appointments = Appointment.listFromJson(JsonUtils.decodeList(appointmentsJsonString));
+      if (_appointments != null) {
+        _sortAppointments(_appointments);
+        _saveAppointmentsStringToCache(appointmentsJsonString);
+      }
+    }
+
     await _initAccount();
     await super.initService();
   }
@@ -94,65 +119,85 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
     }
   }
 
-  // APIs
-  Future<List<Appointment>?> loadAppointments({bool? onlyUpcoming, AppointmentType? type}) async {
-    List<Appointment>? appointments;
-   appointments = await loadBBAppointments();
-    List<Appointment>? resultAppointments;
-    //TBD: Appointment - do this filtering on the backend
-    if (CollectionUtils.isNotEmpty(appointments)) {
-      resultAppointments = <Appointment>[];
-      for (Appointment appointment in appointments!) {
-        if ((onlyUpcoming == true) && !appointment.isUpcoming) {
-          continue;
-        }
-        if ((type != null) && (type != appointment.type)) {
-          continue;
-        }
-        resultAppointments.add(appointment);
-      }
-      _sortAppointments(resultAppointments);
-    }
-    return resultAppointments;
-  }
-
-  Future<List<Appointment>?> loadBBAppointments() async {
-    if (StringUtils.isEmpty(Config().appointmentsUrl)) {
-      Log.w('Failed to appointments. Missing appointments url.');
-      return null;
-    }
-    String? url;
-    url = "${Config().appointmentsUrl}/services/appointments";
-    http.Response? response = await Network().get(url, auth: Auth2());
-    int? responseCode = response?.statusCode;
-    String? responseString = response?.body;
-    if (responseCode == 200) {
-      List<Appointment>? items = Appointment.listFromJson(JsonUtils.decodeList(responseString));
-      return items;
-    } else {
-      Log.w('Failed to load Appointments. Response:\n$responseCode: $responseString');
-      return null;
-    }
-  }
-
-  //TBD remove when not needed for testing
-  Future<List<Appointment>?> loadAssetsAppointments() async {
-    return Appointment.listFromJson(JsonUtils.decodeList(await AppBundle.loadString('assets/appointments.json')));
-  }
-
-  //TBD: Appointment - load from backend
   Future<Appointment?> loadAppointment(String? appointmentId) async {
-    if (StringUtils.isNotEmpty(appointmentId)) {
-      List<Appointment>? allAppointments = await loadAppointments();
-      if (CollectionUtils.isNotEmpty(allAppointments)) {
-        for (Appointment appointment in allAppointments!) {
-          if (appointment.id == appointmentId) {
-            return appointment;
-          }
+    if (StringUtils.isNotEmpty(appointmentId) && StringUtils.isNotEmpty(Config().appointmentsUrl)) {
+      if (StringUtils.isNotEmpty(Config().appointmentsUrl)) {
+        String? url = "${Config().appointmentsUrl}/services/appointments?_ids=$appointmentId";
+        http.Response? response = await Network().get(url, auth: Auth2());
+        int? responseCode = response?.statusCode;
+        String? responseString = response?.body;
+        if (responseCode == 200) {
+          return Appointment.listFromJson(JsonUtils.decodeList(responseString))?.first;
+        } else {
+          debugPrint('Failed to load appointment with id {$appointmentId}. Reason: $responseCode, $responseString');
+          return null;
         }
       }
     }
     return null;
+  }
+
+  List<Appointment>? getAppointments({bool? onlyUpcoming, AppointmentType? type}) {
+    List<Appointment>? result;
+    if (CollectionUtils.isNotEmpty(_appointments)) {
+      result = <Appointment>[];
+      for (Appointment appt in _appointments!) {
+        if ((onlyUpcoming == true) && !appt.isUpcoming) {
+          continue;
+        }
+        if ((type != null) && (type != appt.type)) {
+          continue;
+        }
+        result.add(appt);
+      }
+    }
+    return result;
+  }
+
+  Future<void> refreshAppointments() => _updateAppointments();
+
+  File _getAppointmentsCacheFile() => File(join(_appDocDir.path, _appointmentsCacheDocName));
+
+  Future<String?> _loadAppointmentsStringFromCache() async {
+    File appointmentsFile = _getAppointmentsCacheFile();
+    return await appointmentsFile.exists() ? await appointmentsFile.readAsString() : null;
+  }
+
+  Future<void> _saveAppointmentsStringToCache(String? value) async {
+    await _getAppointmentsCacheFile().writeAsString(value ?? '', flush: true);
+  }
+
+  Future<List<Appointment>?> _loadAppointmentsFromCache() async {
+    return Appointment.listFromJson(JsonUtils.decodeList(await _loadAppointmentsStringFromCache()));
+  }
+
+  Future<String?> _loadAppointmentsStringFromNet() async {
+    //TMP: assets shortcut
+    //return await AppBundle.loadString('assets/appointments.json')
+    if (StringUtils.isNotEmpty(Config().appointmentsUrl)) {
+      String? url = "${Config().appointmentsUrl}/services/appointments";
+      http.Response? response = await Network().get(url, auth: Auth2());
+      int? responseCode = response?.statusCode;
+      String? responseString = response?.body;
+      if (responseCode == 200) {
+        return responseString;
+      } else {
+        debugPrint('Failed to load appointments from net. Reason: $responseCode, $responseString');
+        return null;
+      }
+    }
+    return null;
+  }
+
+  Future<void> _updateAppointments() async {
+    String? appointmentsJsonString = await _loadAppointmentsStringFromNet();
+    List<Appointment>? appointments = Appointment.listFromJson(JsonUtils.decodeList(appointmentsJsonString));
+    if (!DeepCollectionEquality().equals(_appointments, appointments)) {
+      _appointments = appointments;
+      _sortAppointments(_appointments);
+      _saveAppointmentsStringToCache(appointmentsJsonString);
+      NotificationService().notify(notifyAppointmentsChanged);
+    }
   }
 
   void _sortAppointments(List<Appointment>? appointments) {
@@ -193,7 +238,7 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
       _isLastAccountResponseSuccessful = null;
       _account = null;
     }
-    NotificationService().notify(notifyAppointmentsAccountUpdated);
+    NotificationService().notify(notifyAppointmentsAccountChanged);
   }
 
   Future<void> _initAccountOnFirstSignIn() async {
@@ -214,7 +259,7 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
       _isLastAccountResponseSuccessful = false;
       _account = null;
     }
-    NotificationService().notify(notifyAppointmentsAccountUpdated);
+    NotificationService().notify(notifyAppointmentsAccountChanged);
   }
 
   void changeAccountPreferences({bool? newAppointment, bool? morningReminder, bool? nightReminder}) {
@@ -253,7 +298,7 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
       _isLastAccountResponseSuccessful = false;
       _account = null;
     }
-    NotificationService().notify(notifyAppointmentsAccountUpdated);
+    NotificationService().notify(notifyAppointmentsAccountChanged);
   }
 
   // ExploreJsonHandler
@@ -317,6 +362,7 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
       _onAppLivecycleStateChanged(param);
     } else if (name == Auth2.notifyLoginChanged) {
       _initAccount();
+      _updateAppointments();
     }
   }
 
@@ -329,6 +375,7 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
         Duration pausedDuration = DateTime.now().difference(_pausedDateTime!);
         if (Config().refreshTimeout < pausedDuration.inSeconds) {
           _loadAccount();
+          _updateAppointments();
         }
       }
     }